requirements.txt
awsebcli==3.20.3
beautifulsoup4==4.9.3
botocore==1.23.54
cement==2.8.2
certifi==2022.5.18.1
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.3
cssselect==1.1.0
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.7.1
Flask==2.0.3
future==0.16.0
googlesearch-python==1.1.0
idna==2.10
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
jmespath==0.10.0
joblib==1.1.0
lxml==4.9.0
MarkupSafe==2.1.1
newspaper3k==0.2.8
nltk==3.7
numpy==1.22.4
pandas==1.4.2
pathspec==0.9.0
Pillow==9.1.1
pypiwin32==223
python-dateutil==2.8.2
pytz==2022.1
pywin32==304
PyYAML==5.4.1
regex==2022.6.2
requests==2.25.1
requests-file==1.5.1
semantic-version==2.8.5
sgmllib3k==1.0.0
six==1.14.0
soupsieve==2.3.2.post1
tabulate==0.8.9
termcolor==1.1.0
tinysegmenter==0.3
tldextract==3.3.0
tqdm==4.64.0
urllib3==1.26.9
wcwidth==0.1.9
Werkzeug==2.1.2
